Add files via upload
Objectives to understand that AdaBoost is a linear combination of 𝑇 weak classifiers

Understand Hyperparameters selection in AdaBoost Like Bagging and Random Forest (RF), AdaBoost combines the outputs of many classifiers into an ensemble, but there are some differences. Each classifier in the ensemble is powerful but prone to overfitting. As Bagging or RF aggregate more and more classifiers, they reduce overfitting.